% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hellno.R
\name{data.frame}
\alias{data.frame}
\title{alternative data.frame() implementation}
\usage{
data.frame(..., row.names = NULL, check.rows = FALSE, check.names = TRUE,
  stringsAsFactors = HELLNO)
}
\description{
\code{hellno::data.frame()} wraps up \code{base::data.frame()} so that
\code{stringAsFactors} is set to \code{HELLNO} ( \code{== FALSE} ) by default
}
\seealso{
\code{\link[base]{data.frame}}
}

