% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellno.R
\name{rbind.data.frame}
\alias{rbind.data.frame}
\title{altenative rbind.data.frame() implementation}
\usage{
rbind.data.frame(..., deparse.level = 1, make.row.names = TRUE,
  stringsAsFactors = HELLNO)
}
\arguments{
\item{...}{see \code{\link[base]{rbind.data.frame}}}

\item{deparse.level}{see \code{\link[base]{rbind.data.frame}}}

\item{make.row.names}{see \code{\link[base]{rbind.data.frame}}}

\item{stringsAsFactors}{\code{\link[base]{rbind.data.frame}} by default set to \code{FALSE}}
}
\description{
altenative rbind.data.frame() implementation
}
\seealso{
\code{\link[base]{rbind.data.frame}}
}

